Give alarm-server a low oom_adj value to avoid killed
Currently, a positive oom_adj make alarm-server likely to be killed by
system-server, which will pick up the one with positive oom_adj and
increase it's oom_adj whenever it go to inactive or backgrd. Once its
oom_adj larger or equal 15, it got killed by system-server.

For sysvinit, there is a kernel module named pmon and user space tool
which set important services as vip or permenant processes. So they're
lesser to be killed and will be relaunched once it got killed.

In systemd environment, it's easy to implement this feature by leverage
the systemd feature.

Change-Id: Ib89cdd1c94cfac66e1da88144db98ec1314b8865
Signed-off-by: Chengwei Yang <chengwei.yang@intel.com>